+++ package.json
@@ -18,6 +18,7 @@
   },
   "homepage": "https://starb.in",
   "devDependencies": {
+    "@cloudflare/workers-types": "^3.10.0",
     "prettier": "^2.2.1",
     "wrangler": "^2.0.5"
   }
